use trip_instance_key, remove sched_rt_util functions with updated columns in warehouse

+++ open_data/concatenate_vehicle_positions.py
@@ -42,7 +42,7 @@ def concat_batches(analysis_date: str) -> dd.DataFrame:
         ddf, ["location_timestamp"])
         
     ddf = ddf.assign(
-        hour = dd.to_datetime(ddf.location_timestamp_local).dt.hour,
+        hour = dd.to_datetime(ddf.location_timestamp_local).dt.hour.astype("int8"),
     )
     
     return ddf
